Migrate setup.py to pyproject.toml (#8)
