+++ setup.cfg
@@ -1,0 +1,3 @@
+[metadata]
+url = "https://github.com/thoth-station/termial-random"
+download_url = "https://pypi.org/project/termial-random"
